arm64: DT: beryllium: Explicitly write smb1355 nodes inside
Instead of including the smb1355.dtsi file, write the nodes in here.
This solution was chosen because the k4.9 smb1355 dtsi does
conflict with the one in 4.14, used by other SoCs which we
may be interested in.
Copying the 4.9 smb1355.dtsi as smb1355-845.dtsi that would be
obviously used _only_ by SDM845 would be just adding one more
file without having any sense.
Let's at least keep the filecount lower...